OneDrive/Desktop/cloud-notes-main/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
PyMongo==4.3.3
bson==0.5.10
gunicorn==21.2.0
python-dotenv==1.0.0
